% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_stan_dist_int}
\alias{assign_stan_dist_int}
\title{Return the integer value for STAN that corresponds to each supported distribution}
\usage{
assign_stan_dist_int(distribution)
}
\arguments{
\item{distribution}{specified distribution}
}
\value{
Integer value corresponding to chosen distribution
}
\description{
Currently supported distributions: normal (0), uniform (1), exponential (2).
}
\details{
Used internally, in \code{\link{prep_init_list}}.
}
\keyword{internal}
