% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_init_chain}
\alias{check_init_chain}
\title{Check whether the random initial values for a chain are appropriate.}
\usage{
check_init_chain(single.init.list)
}
\arguments{
\item{single.init.list}{The list of initial values to be checked.}
}
\value{
NULL, if no problems, otherwise a vector of parameters which have
inappropriate initial values.
}
\description{
Checks the initial values before they are passed to Stan to make sure they are appropriate.
}
\details{
Used internally, in \code{\link{prep_init_list}}.

Checks that: width is positive, delta parameters are positive, tau
parameters, pmin/pmax, and f are between 0 and 1, and that pmin is less than
pmax.
}
\examples{
\dontrun{
assign_stan_dist_int("normal") # returns 0
assign_stan_dist_int("uniform") # returns 1
assign_stan_dist_int("exponential") # returns 2
}

}
\keyword{internal}
