% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prior_specification}
\alias{check_prior_specification}
\title{Check that the distribution specified has the right number of parameters}
\usage{
check_prior_specification(distribution, string)
}
\arguments{
\item{distribution}{the prior distribution, as a character string.}

\item{string}{the string specifiying the parameter values, parsed from the prior_config file.}
}
\value{
TRUE/FALSE: is the prior specified correctly?
}
\description{
Used internally, in \code{\link{prep_prior_list}}.
}
\examples{
\dontrun{
check_prior_specification("normal", "normal(32,45)") # returns T
check_prior_specification("uniform", "uniform(32)") # returns F
check_prior_specification("exponential", "exponential(0.5)") # returns T
}

}
\keyword{internal}
