% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prior_supported}
\alias{check_prior_supported}
\title{Check that the distribution specified for a given parameter is supported}
\usage{
check_prior_supported(parameter, distribution)
}
\arguments{
\item{parameter}{the parameter for which a prior distribution is specified}

\item{distribution}{the prior distribution for that parameter}
}
\value{
TRUE/FALSE: is the prior supported for that parameter?
}
\description{
Used internally, in \code{\link{make_prior_list}}.
}
\keyword{internal}
