% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_bahz_cache.R
\name{clear_bahz_cache}
\alias{clear_bahz_cache}
\title{Clear cache of saved results from memoised functions in bahz}
\usage{
clear_bahz_cache()
}
\value{
invisible(NULL)
}
\description{
Clear the cache of all already calculated results for all the memoised functions
(predict_cline, cline_summary) in bahz.
}
\examples{
\dontrun{

clear_bahz_cache()

}

}
