% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cline_summary.R
\name{cline_summary}
\alias{cline_summary}
\title{Summarize cline model results}
\usage{
cline_summary(stanfit, prob = 0.95, method = "ET", all = F)
}
\arguments{
\item{stanfit}{A stanfit object holding your model results.}

\item{prob}{The probability interval to return. Default is .95. Numeric,
between 0 and 1.}

\item{method}{The method for calculating credible intervals. Either "ET" for
equal-tail probability intervals, or "HPDI" for highest posterior density
intervals.}

\item{all}{By default, function returns summaries only for the main cline
parameters. Set all to T}
}
\value{
A summary data frame with the columns:
\itemize{
\item param:
\item mean:
\item se_mean:
\item sd:
\item lower:
\item upper:
\item n_eff:
\item Rhat:
}
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED. LINK TO RSTAN SUMMARY METHOD, CODA HPDI, MAYBE SOME
PAPERS ON WHAT THE THINGS ARE.
}
\examples{
# TO ADD

}
