% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_fis}
\alias{correct_fis}
\title{Correct Fis estimates}
\usage{
correct_fis(.df)
}
\arguments{
\item{.df}{A data frame, containing a column named Fis.
Meant to be run as part of a dplyr pipeline}
}
\value{
TO ADD
}
\description{
A quick little function to turn NaNs and negative numbers into 0
}
\details{
DETAILS TO BE ADDED. LINK TO THE GENERAL CLINE EQUATION FUNCTION, MAYBE MOVE
THE CLINE PARAMETER DEFINITIONS THERE.
}
\examples{
# to be added



}
\keyword{internal}
