% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cline_model.R
\name{create_cline_model}
\alias{create_cline_model}
\title{Create the model code for a cline analysis}
\usage{
create_cline_model(prior_file, type = c("bi", "multi"),
  tails = c("none", "left", "right", "mirror", "ind"),
  direction = c("inc", "dec"))
}
\arguments{
\item{prior_file}{The path to the \code{.yaml} file which contains the
specifications of the priors}

\item{type}{TO BE WRITTEN}

\item{tails}{TO BE WRITTEN}

\item{direction}{TO BE WRITTEN}
}
\value{
TO BE WRITTEN
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
#TO BE ADDED

}
