% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cline_models.R
\name{create_cline_models}
\alias{create_cline_models}
\title{Create the cline models}
\usage{
create_cline_models(prior_file)
}
\arguments{
\item{prior_file}{The path to the \code{.yaml} file which contains the
specifications of the priors}
}
\value{
TO BE WRITTEN
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
#TO BE ADDED

}
