% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_first}
\alias{extract_first}
\title{Extract the first value from comma-separated list/vector}
\usage{
extract_first(string)
}
\arguments{
\item{string}{The string to extract a value from}
}
\value{
A numeric value
}
\description{
Uses regular expressions implmented in stringr package to extract the first
value, that is, the one after an open parenthesis and before a comma.
Properly handles whitespace and decimals.
}
\details{
Used interanlly, in \link{init_single_chain}
}
\keyword{internal}
