% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_last}
\alias{extract_last}
\title{Extract the last value from comma-separated list/vector}
\usage{
extract_last(string)
}
\arguments{
\item{string}{The string to extract a value from}
}
\value{
A numeric value
}
\description{
Uses regular expressions implmented in stringr package to extract the last
value, that is, the one after a comma and before a close parenthesis.
Properly handles whitespace and decimals.
}
\details{
Used interanlly, in \link{init_single_chain}
}
\keyword{internal}
