% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_only}
\alias{extract_only}
\title{Extract the value from a single-parameter distribution}
\usage{
extract_only(string)
}
\arguments{
\item{string}{The string to extract a value from}
}
\value{
A numeric value
}
\description{
Uses regular expressions implmented in stringr package to extract the only
value from a distribution with one parameter. That is, is extracts the
numbers between two parentheses. Properly handles whitespace and decimals.
}
\details{
Used interanlly, in \link{init_single_chain}
}
\keyword{internal}
