% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extractValue}
\alias{extractValue}
\alias{extract_first}
\alias{extract_last}
\alias{extract_only}
\title{Extract values for prior distributions}
\usage{
extract_first(string)

extract_last(string)

extract_only(string)
}
\arguments{
\item{string}{The string to extract a value from.}
}
\value{
A numeric value.
}
\description{
Internal functions used in \code{\link{prep_init_list}}
and \code{\link{prep_prior_list}} to extract
numerical values from priors. All use regular expressions as implemented
in the \code{\link{stringr}} package, and all properly handle whitespace and
decimals.

\code{extract_first} gets the first value, that is, the one after an open
parenthesis and before a comma.

\code{extract_last} gets the last value, that is, the value after a comma and
before a close parenthesis.

\code{extract_only} gets the only value from a distribution with one
parameter, that is, it gets the numbers between two parentheses.
}
\examples{
\dontrun{
extract_first("uniform(32,45)") # returns 32
extract_last("uniform(32,45)") # returns 45
extract_only("exponential(32)") # returns 32
}
}
\keyword{internal}
