% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_geno_cline.R
\name{fit_geno_cline}
\alias{fit_geno_cline}
\title{Fit a genetic cline model to your data}
\usage{
fit_geno_cline(data, prior_file, type = c("bi", "multi"),
  tails = c("none", "left", "right", "mirror", "ind"), chains = 3, ...)
}
\arguments{
\item{data}{A dataframe containing your cline data. See
\code{\link{prep_geno_data}} for possible formats.}

\item{prior_file}{The path to the \code{.yaml} file which contains the
specifications of the priors.}

\item{type}{The type of model to fit Either "bi", for a binomial model
of allele frequencies, or "multi" for a multinomial model of genotype
frequencies.}

\item{tails}{Which type of tails for the model: "none", "left", "right", "mirror", or
"ind"?}

\item{chains}{The number of MCMC chains to create. Numeric, coerced to
integer. Default is 3.}

\item{...}{Arguments to be passed to \code{stan}, e.g., number of iterations, warmup
period, etc. See \code{\link[rstan]{sampling}} for more information.}
}
\value{
A \code{\linkS4class{stanfit}} object containing your model results.
}
\description{
Use Rstan to fit Bayesian hybrid zone cline models to genetic data.
}
\details{
This is a wrapper function, which calls various data- and model-preparation
functions from \code{bahz} before passing the results to the
\code{\link[rstan]{sampling}} function in \code{rstan}, which does the MCMC model
fitting. Specifically, this function calls:

\enumerate{
\item \code{\link{prep_geno_data}} to format the provided
\code{data}.
\item \code{\link{prep_prior_list}}, which creates a list containing
the specifications of the prior distributions, to be passed to Stan.
It creates the priors according to the configuration file \code{prior_file}.
\item \code{\link{prep_init_list}}, which creates a list of values for
initialiing the MCMC chains in \code{stan}. Initial values are randomly
drawn from the priors specified in \code{prior_file}.
}

The results of these functions are then passed to \code{\link[rstan]{sampling}}, along with
any of the additional \code{stan} arguments supplied by the user.
}
\examples{
\dontrun{
# Fit a multinomial cline
# with mirrored introgression tails.
# Uses default number of chains and stan parameters.
results <- fit_geno_cline(clinedata, "prior_file.yaml",
                     type = "multi", tails = "mirror")

# Fit a binomial cline
# with no introgression tails.
# Use 4 chains with 4000 warmup iterations
# and 8000 total iterations per chain.

results2 <- fit_geno_cline(clinedata, "prior_file.yaml",
                     type = "bi", tails = "none",
                     chains = 4,
                     iter = 8000, warmup = 4000)
}




}
\seealso{
\code{\link{prep_geno_data}}, \code{\link{prep_prior_list}},
\code{\link{prep_init_list}}, \code{\link[rstan]{sampling}}
}
