% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cline_eqn.R
\name{general_cline_eqn}
\alias{general_cline_eqn}
\title{Calculate values from cline equations}
\usage{
general_cline_eqn(transectDist, decrease = F, center, width, pmin = 0,
  pmax = 1, deltaL = NULL, tauL = NULL, deltaR = NULL,
  tauR = NULL)
}
\arguments{
\item{transectDist}{The distances along the transect for the sampling
sites to be simulated. A numeric vector.}

\item{decrease}{Is the cline decreasing in frequency? \code{TRUE} or
\code{FALSE} (Default = F).}

\item{center}{The location of the cline center, in the same distance units as
\code{transect_distances}. Numeric.}

\item{width}{The width of the cline, in the same distance units as
\code{transect_distances}. Numeric.}

\item{pmin, pmax}{The minimum and maximum allele frequency values across the
cline. Must be between 0 and 1 (inclusive). Numeric.}

\item{deltaL, tauL}{TO BE ADDED}

\item{deltaR, tauR}{TO BE ADDED}
}
\value{
TO ADD
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
# to be added


}
