% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cline_eqn.R
\name{internal_cline_eqn}
\alias{internal_cline_eqn}
\title{Calculate allele frequencies for a single point along a cline}
\usage{
internal_cline_eqn(transectDist, decrease = c(TRUE, FALSE), center,
  width, pmin = 0, pmax = 1, deltaL = NA, tauL = NA, deltaR = NA,
  tauR = NA)
}
\arguments{
\item{transectDist}{The position at which to evaluate the cline equation.
Must be a numeric vector.}

\item{decrease}{Is the cline decreasing in frequency? \code{TRUE} or
\code{FALSE}.}

\item{center}{The location of the cline center, in the same distance units as
\code{transectDist}. Numeric.}

\item{width}{The width of the cline, in the same distance units as
\code{transectDist}. Numeric, must be greater than 0.}

\item{pmin, pmax}{Optional. The minimum and maximum allele frequency values in
the tails of the cline. Default values are \code{0} and \code{1},
respectively. Must be between 0 and 1 (inclusive). Numeric.}

\item{deltaL, tauL}{Optional delta and tau parameters which describe the left
exponential introgression tail. Must supply both to generate a tail. Default is
\code{NA} (no tails). Numeric. tauL must be between 0 and 1 (inclusive).}

\item{deltaR, tauR}{Optional delta and tau parameters which describe the right
exponential introgression tail. Must supply both to generate a tail. Default is
\code{NA} (no tails). Numeric. tauR must be between 0 and 1 (inclusive).}
}
\value{
The result of evaluating a cline equation with the specified
parameters at the specified distance (that is, \eqn{p'} in the notation of the
equations above). A numeric vector of length 1.
}
\description{
A flexible function to evaluate hybrid zone genetic cline equations at a
single point. Used internally by \code{\link{general_cline_eqn}}, which
calculates results for multiple points. See that help page for all the math
details.
}
\keyword{internal}
