% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_cline_data.R
\name{load_cline_data}
\alias{load_cline_data}
\title{Prepare your cline data for loading into Stan}
\usage{
load_cline_data(dataframe, type)
}
\arguments{
\item{dataframe}{A dataframe containing your cline data. See details for
possible formats.}

\item{type}{Model type. Either "bi", for the binomial model, or "multi", for
the multinomial model.}
}
\value{
Your data, in the list format appropriate for the cline model you
chose.
}
\description{
DESCRIPTION TO BE WRITTEN
}
\details{
DETAILS TO BE ADDED. SPECIFICALLY, REQUIRED FILE FORMAT.
}
\examples{
# TO BE ADDED

}
