% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_init_list.R
\name{make_init_list}
\alias{make_init_list}
\title{Make a list of vlaues for initializing the stan model.}
\usage{
make_init_list(prior_file, tails, chains)
}
\arguments{
\item{prior_file}{Path to the yaml file containing the prior specifications.}

\item{tails}{Which type of tails for the model: "none", "left", "right", "mirror", or "ind."}

\item{chains}{The number of chains to be used in Stan.}
}
\value{
A list of lists containing initialization values for the Stan model.
}
\description{
This function calls \link{init_single_chain}, which does all the work of
making the expression which will yeild the proper starting values for a chain
of this model. It then evaluates that expression for each desired chain in
the model and adds it to a list.
}
\details{
Currently supported distributions are: normal, uniform, exponential.
}
