% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_config.R
\name{make_prior_config}
\alias{make_prior_config}
\title{Make a prior configuration file}
\usage{
make_prior_config(path = getwd(), name = "prior_config_template.yaml",
  overwrite = F)
}
\arguments{
\item{path}{The folder where you wish to save the prior configuration file.
The default is the current working directory.}

\item{name}{The name you want to give the prior configuration file. The
default is \code{prior_config_template.yaml}. File name must end in \code{.yaml}.}

\item{overwrite}{Do you want to overwrite existing files?
\code{TRUE} or \code{FALSE}, default is \code{FALSE}.}
}
\value{
\code{invisible(NULL)}
}
\description{
Makes a configuration file which you can use to specify the priors for your
hybrid zone analysis. By default, creates the file in the working directory
with the name \code{prior_config_template.yaml}. You can supply a filepath
and a different name, if you want to save the file somewhere else or with a
different name. Will not overwrite files by default, you must set overwrite =
T to do that.
}
\examples{
\dontrun{
# Make config file in current working directory with default name.
# Won't overwrite
make_prior_config()

# Make config file in a different folder with a new name.
# Will overwrite an existing file.
make_prior_config(path = "/path/to/directory", name = "my_priors.yaml", overwrite = T)
}

}
