% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_prior_list.R
\name{make_prior_list}
\alias{make_prior_list}
\title{Make the list of prior value to be fed into Stan}
\usage{
make_prior_list(prior_file)
}
\arguments{
\item{prior_file}{TO BE WRITTEN}
}
\value{
A list containing the values for the priors to be passed to Stan
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
#TO BE ADDED

}
