% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_prior_file}
\alias{parse_prior_file}
\title{Parse the prior yaml file and check it}
\usage{
parse_prior_file(prior_file)
}
\arguments{
\item{prior_file}{Filepath to the prior file.}
}
\value{
A named list containing the specifed priors.
}
\description{
Reads in the yaml file containing the prior, checking for the proper number
and names of priors.
}
\details{
Used internally, in \code{\link{prep_prior_list}} and
\code{\link{prep_init_list}}.
}
\examples{
\dontrun{
parse_prior_file(path/to/priors.yaml)
}

}
\keyword{internal}
