% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cline.R
\name{plot_cline}
\alias{plot_cline}
\title{Pot thing}
\usage{
plot_cline(stanfit)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model results.}
}
\value{
a plot
}
\description{
Pot thing
}
\examples{
\dontrun{
plot_cline(yourStanfit)
}
}
