% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_geno_cline.R
\name{predict_geno_cline}
\alias{predict_geno_cline}
\title{Pot thing}
\usage{
predict_geno_cline(stanfit, data, num.out = NULL)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model results.}

\item{data}{The dataframe with your cline data.}

\item{num.out}{Optional, the number of x values (distances) at which to
evaluate the cline. By default, does twice the length of the cline, but you
can specify more or fewer. Too few may leade to a jagged-looking cline.}
}
\value{
a plot
}
\description{
Pot thing
}
\examples{
\dontrun{

predict_genocline(yourStanfit, data)

# If you want to specify only 100 data points for plotting
predict_genocline(yourStanfit, data, num.out = 100)
}
}
