% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_geno_data.R
\name{prep_geno_data}
\alias{prep_geno_data}
\title{Prepare your genetic cline data for loading into Stan}
\usage{
prep_geno_data(dataframe, type)
}
\arguments{
\item{dataframe}{A dataframe containing your cline data. See details for
possible formats.}

\item{type}{Model type. Either "bi", for the binomial model, or "multi", for
the multinomial model.}
}
\value{
Your data, in the list format appropriate for the cline model you
chose.
}
\description{
Converts a dataframe containing your genetic data into the list format required for
input to Stan.
}
\details{
The input dataframe can be a data frame or a tibble. For the multinomial
model, four named columns must be present, all other columns will be
ignored:
\itemize{
\item transectDist: A numeric column, giving the position along
the cline/transect for each site.
\item AA: The number of sampled individuals that are homozygous for the
focal allele. Integer.
\item Aa: The number of sampled individuals that are heterozygotes.
Integer.
\item aa: The number of sampled individuals that are homozygous for the
non-focal allele. Integer.
}

For the binomial model, the user can supply either a data frame with the four
columns above, or a dataframe with three named columns present (again, other
columns are ignored):

\itemize{
\item transectDist: A numeric column, giving the position along
the cline/transect for each site.
\item nFocalAllele: The allele count for the focal allele.
\item nTotalAlleles: The total number of alleles sampled (twice the
number of diploid individiduals).
}
}
\examples{
\dontrun{
prep_geno_data(yourdata, type = "bi")
}

}
\seealso{
\code{\link{fit_geno_cline}}
}
