% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_init_list.R
\name{prep_init_list}
\alias{prep_init_list}
\title{Prepare the list of values for initializing the Stan model}
\usage{
prep_init_list(prior_file, tails = c("none", "left", "right", "mirror",
  "ind"), type = c("bi", "multi"), chains)
}
\arguments{
\item{prior_file}{Path to the \code{.yaml} file containing the prior specifications.}

\item{tails}{Which type of tails for the model: "none", "left", "right", "mirror", or "ind"?}

\item{type}{The type of model to fit. Either "bi", for a binomial model
of allele frequencies, or "multi" for a multinomial model of genotype
frequencies.}

\item{chains}{The number of chains to be used in Stan. Integer.}
}
\value{
A list of lists containing initialization values for the Stan model.
}
\description{
Generates a list, of length \code{chains}, where each element of the list is
a named list giving the initialization values for each model parameter.
Initialization values are randomly drawn from the prior distribution, and are
drawn independently for each chain.
}
\details{
Currently supported distributions are: normal, uniform, exponential.

This function calls \code{\link{init_single_chain}} for each chain.
\code{\link{init_single_chain}} does all the work of making the expression
which will yield the proper starting values for a chain of this model.

This function also provides checks that the specified starting values will
work with Stan. If the priors are poorly specified such that some initial
values are inappropriate (e.g., width initial values are negative), it will
re-do the initialization, up to 100 times per chain. If, after 100 tries, it fails to
generate an appropriate starting chain, it will generate a warning and
provide some guidance on which parameters are causing the problem.
}
\examples{
\dontrun{
prep_init_list("path/to/priors.yaml", tails = "none", chains = as.integer(3))
}
}
\seealso{
\code{\link{init_single_chain}}, \code{\link{fit_geno_cline}}
}
