% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_pheno_data.R
\name{prep_pheno_data}
\alias{prep_pheno_data}
\title{Prepare your phenotypic cline data for loading into Stan}
\usage{
prep_pheno_data(dataframe)
}
\arguments{
\item{dataframe}{A dataframe containing your phenotypic cline data. See
details for the necessary format.}
}
\value{
Your data in Stan-ready list format.
}
\description{
Converts a dataframe containing your phenotypic data into the list format required for
input to Stan. See details below for a description of the possible input formats.
}
\details{
The input dataframe can be a data frame or a tibble. Each row should contain
the information for a single sampled individual. Two named
columns must be present, all other columns will be ignored:
\itemize{
\item transectDist: A numeric column, giving the position along the
cline/transect for each individual.
\item traitValue: A numeric column
giving the phenotypic trait value for each individual.
}
}
\examples{
\dontrun{
prep_pheno_data(yourdata)
}

}
\seealso{
\code{\link{fit_pheno_cline}}
}
