% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_prior_list.R
\name{prep_prior_list}
\alias{prep_prior_list}
\title{Prepare the list of prior values to be fed into Stan}
\usage{
prep_prior_list(prior_file)
}
\arguments{
\item{prior_file}{Path to the yaml file containing the prior specifications.}
}
\value{
A list containing the values for the priors to be passed to Stan
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
#TO BE ADDED
}
