% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_from_cline.R
\name{sim_data_from_cline}
\alias{sim_data_from_cline}
\title{Simulate genetic data from a cline}
\usage{
sim_data_from_cline(transect_distances, n_ind, Fis, decrease = F, center,
  width, pmin = 0, pmax = 1, deltaL = NULL, tauL = NULL,
  deltaR = NULL, tauR = NULL)
}
\arguments{
\item{transect_distances}{The distances along the transect for the sampling
sites to be simulated. A numeric vector.}

\item{n_ind}{The number of diploid individuals sampled at each site. Either a
single numeric value (for constant sampling), or a numeric vector equal in
length to \code{transect_distances}.}

\item{Fis}{The inbreeding coefficient, Fis, for each site. Must be between 0
and 1 (inclusive). Either a single numeric value (for constant inbreeding),
or a numeric vector equal in length to \code{transect_distances}.}

\item{decrease}{Is the cline decreasing in frequency? \code{TRUE} or
\code{FALSE} (Default = F).}

\item{center}{The location of the cline center, in the same distance units as
\code{transect_distances}. Numeric.}

\item{width}{The width of the cline, in the same distance units as
\code{transect_distances}. Numeric.}

\item{pmin}{The minimum and maximum allele frequency values across the
cline. Must be between 0 and 1 (inclusive). Numeric.}

\item{pmax}{The minimum and maximum allele frequency values across the
cline. Must be between 0 and 1 (inclusive). Numeric.}

\item{deltaL}{TO BE ADDED}

\item{tauL}{TO BE ADDED}

\item{deltaR}{TO BE ADDED}

\item{tauR}{TO BE ADDED}
}
\value{
TO ADD
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED. LINK TO THE GENERAL CLINE EQUATION FUNCTION, MAYBE MOVE
THE CLINE PARAMETER DEFINITIONS THERE.
}
\examples{
# to be added


}
