% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_geno_cline.R
\name{sim_geno_cline}
\alias{sim_geno_cline}
\title{Simulate genetic data from a cline}
\usage{
sim_geno_cline(transect_distances, n_ind, Fis, decrease, center, width,
  pmin = 0, pmax = 1, deltaL = NA, tauL = NA, deltaR = NA,
  tauR = NA)
}
\arguments{
\item{transect_distances}{The distances along the transect for the simulated
sampling sites. A numeric vector.}

\item{n_ind}{The number of diploid individuals sampled at each site. Either a
single numeric value (for constant sampling), or a numeric vector equal in
length to \code{transect_distances}.}

\item{Fis}{The inbreeding coefficient, Fis, for each site. Must be between 0
and 1 (inclusive). Either a single numeric value (for constant inbreeding),
or a numeric vector equal in length to \code{transect_distances}.}

\item{decrease}{Is the cline decreasing in frequency? \code{TRUE} or
\code{FALSE}.}

\item{center}{The location of the cline center, in the same distance units as
\code{transect_distances}. Numeric, must be greater than 0.}

\item{width}{The width of the cline, in the same distance units as
\code{transect_distances}. Numeric, must be greater than 0.}

\item{pmin, pmax}{Optional. The minimum and maximum allele frequency values
in the tails of the cline. Default values are \code{0} and \code{1}, respectively.
Must be between 0 and 1 (inclusive). Numeric.}

\item{deltaL, tauL}{Optional delta and tau parameters which describe the left
exponential tail. Must supply both to generate a tail. Default is \code{NA} (no
tails). Numeric. tauL must be between 0 and 1 (inclusive).}

\item{deltaR, tauR}{Optional delta and tau parameters which describe the right
exponential tail. Must supply both to generate a tail. Default is \code{NA} (no
tails). Numeric. tauR must be between 0 and 1 (inclusive).}
}
\value{
A data frame of simulated genetic data sampled from the cline. Columns are:
\itemize{
\item site: The site numbers, given sequentially starting at 1.
\item transectDist: The distance along the cline for each site.
\item cline.p: The expected allele frequency for each site, given its position on the cline.
\item cline.f: The expected coefficient of inbreeding for each site.
\item AA, Aa, aa: The simulated number of homozygotes and heterozygotes for each site.
\item N: The number of individuals sampled for each site.
\item emp.p: The observed allele frequency for each site (includes sampling error).
\item emp.f: The observed Fis for each site (includes sampling error).
}
}
\description{
This function generates a dataframe with simulated genotypic data sampled
from a genetic cline. The sampling sites, number of individuals, level of
inbreeding, and cline parameters are supplied by the user. Cline parameters
are flexible, and can model both sigmoid clines and stepped clines with
introgresison tails.
}
\details{
This function calls \code{\link{general_cline_eqn}} for each sampled point along the
cline to generate the expected allele frequency at that point. The calculated
allele frequency and provided Fis are then used to calculate the expected
genotype frequencies for each site, according to the equations:

\deqn{AA = p^2 + p(1-p)Fis}
\deqn{Aa = 2p(1-p)(1-Fis)}
\deqn{aa = (1-p)^2 + p(1-p)Fis}

Sampled genotypes are then drawn from a multinomial distribution, with the
expected genotype frequencies as the probabilities. From those sampled
genotypes, empirical allele frequencies (emp.p) and empirical Fis estimates
(emp.f) are calculated as:

\deqn{emp.p = (2*AA + Aa)/N}
\deqn{emp.f = (Hexp - Hobs)/Hexp}

where AA is the number of honozygotes of the focal allele, Aa is the number
of heterozygotes, N is the number of sampled individuals, and  Hexp and Hobs
are the expected and observed heterozygosity. Fis values are corrected with
\link{correct_fis}, such that empirical Fis values which are undefined or < 0
are corrected to 0.
}
\examples{
# Simulate genotype data from a decreasing cline
# with center at 100, width of 30.
# Sites are 20 units apart, from 0 to 200.
# 20 individuals are sampled at each site.
# Inbreeding is constant at Fis = 0.1.

set.seed(123)
sim_geno_cline(transect_distance = seq(0,200,20),
               n_ind = 20, Fis = 0.1,
               decrease = TRUE,
               center = 100, width = 30)

# Simulate genotype data from an increasing cline
# with center at 272, width of 91.
# The minimum and maximum allele frequencies
# are 0.07 and 0.98, respectively, and
# there is an introgression tail on the
# left side with deltaL = 29 and tauL = 0.8.

# Sites are 13 units apart, from 162 to 312.
# At each site, the number of individuals sampled
# is drawn from a random normal distribution with
# mean = 25 and sd = 5
# Inbreeding is constant at Fis = 0.

set.seed(123)
ind_sampling <- as.integer(rnorm(length(seq(162,312,12)), 25, 5))
sim_geno_cline(transect_distance = seq(162,312,12),
               n_ind = ind_sampling,
               Fis = 0, decrease = FALSE,
               center = 272, width = 91,
               pmin = 0.07, pmax = 0.98,
               deltaL = 29, tauL = 0.8)

}
