% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit_cline.R
\name{test_fit_cline}
\alias{test_fit_cline}
\title{Fit a cline model to your data using Stan}
\usage{
test_fit_cline(stan_data, init_list, chains, model)
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
