% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_fit_cline2.R
\name{test_fit_cline2}
\alias{test_fit_cline2}
\title{Fit a cline model to your data using Stan}
\usage{
test_fit_cline2(stan_data, init_list, chains)
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
