% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bahz-package.r
\docType{package}
\name{bahz-packge}
\alias{bahz-packge}
\alias{bahz}
\alias{bahz-packge-package}
\title{bahz: Bayesian Analysis of Hybrid Zones}
\description{
A DESCRIPTION OF THE PACKAGE
}
\references{
Stan Development Team (2018). RStan: the R interface to Stan. R package version 2.17.4. http://mc-stan.org
}
