% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prior_specification}
\alias{check_prior_specification}
\title{Check that the distribution specified has the right number of parameters}
\usage{
check_prior_specification(distribution, string)
}
\arguments{
\item{distribution}{the prior distribution}

\item{string}{the string specifiying the parameter values, parsed from the prior_config file}
}
\value{
TRUE/FALSE: is the prior specified correctly?
}
\description{
Used internally, in \code{\link{make_prior_list}}.
}
\keyword{internal}
