% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_prior_supported}
\alias{check_prior_supported}
\title{Check that the distribution specified for a given parameter is supported}
\usage{
check_prior_supported(parameter, distribution)
}
\arguments{
\item{parameter}{The parameter for which a prior distribution is specified,
as a character string.}

\item{distribution}{The prior distribution for that parameter, as a character
string.}
}
\value{
TRUE/FALSE: is the distribution supported for that parameter?
}
\description{
Used internally, in \code{\link{prep_prior_list}}. Checks that the prior
distribution specified for a particular cline parameter is currently
supported.
}
\examples{
\dontrun{
check_prior_supported("center", "normal") # returns T
check_prior_supported("center", "beta") # returns F
check_prior_supported("pmin", "beta") # returns T
check_prior_supported("deltaL", "normal") # returns F
}
}
\keyword{internal}
