% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cline_summary.R
\name{cline_summary}
\alias{cline_summary}
\title{Summarize cline model results}
\usage{
cline_summary(stanfit, prob = 0.95, method = "HPDI", show.all = F)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model results.}

\item{prob}{The probability interval to return. Default is .95. Numeric,
between 0 and 1.}

\item{method}{The method for calculating credible intervals. Either "ET" for
equal-tail probability intervals, or "HPDI" for highest posterior density
intervals. Default is "HPDI".}

\item{show.all}{By default, function returns summaries only for the main cline
parameters. Set show.all to T to see summaries for matrix and vector parameters.}
}
\value{
A summary data frame with the columns:
\itemize{
\item param: The model parameter
\item mean: The mean of the posterior distribution.
\item se_mean: The standard error of the mean of the posterior distribution.
\item sd: The standard deviation of the posterior distribution.
\item lower: The lower limit of the credible interval. The column name
will include the probability value and method used for calculating the
CI.
\item upper: The upper limit of the credible interval. The column name
will include the probability value and method used for calculating the
CI.
\item n_eff: The effective number of samples from the posterior distribution.
\item Rhat: The Gelman-Rubin convergence diagnostic. Should be 1, any
other value is indicative of a problem during sampling.
}
}
\description{
Summarizes the results of the cline model by providing the posterior mean,
credible intervals, and diagnostics for the model parameters.
}
\details{
Uses the \code{rstan} summary method on class \code{\linkS4class{stanfit}} objects
for calculating posterior means, SEM, standard deviations, equal tail
probability intervals, and diagnostics. Uses the \code{\link[coda]{HPDinterval}} function in
\code{coda} to calculate HPDI intervals, the default.
}
\examples{
\dontrun{
cline_summary(stanfit)
}

}
