% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{correct_fis}
\alias{correct_fis}
\title{Correct Fis estimates}
\usage{
correct_fis(.df)
}
\arguments{
\item{.df}{A data frame, containing a column named Fis.}
}
\value{
The supplied data frame, with a corrected Fis column
}
\description{
A function that corrects Fis estimates which are negative or NaN, turns them
to 0. Reads in a data frame with an Fis column. Meant to be part of a dplyr
pipeline.
}
\details{
Used internally, in \code{\link{sim_geno_cline}}.
}
\keyword{internal}
