% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_cline.R
\name{fit_cline}
\alias{fit_cline}
\title{Fit a cline model to your data using Stan}
\usage{
fit_cline(data, prior_file, type, tails, direction, chains = 3, ...)
}
\arguments{
\item{data}{TO BE WRITTEN}

\item{prior_file}{TO BE WRITTEN}

\item{type}{TO BE WRITTEN}

\item{tails}{TO BE WRITTEN}

\item{direction}{TO BE WRITTEN}

\item{chains}{The number of MCMC chains to create. Integer. Default is 3.}

\item{...}{Arguments to be passed to stan. LINK TO IT.}
}
\value{
A stanfit object containing your model results.
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
DETAILS TO BE ADDED.
}
\examples{
#TO BE ADDED

}
