% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_cline_eqn.R
\name{general_cline_eqn}
\alias{general_cline_eqn}
\title{Calculate allele frequency at a position on a cline}
\usage{
general_cline_eqn(transectDist, decrease = c(TRUE, FALSE), center, width,
  pmin = 0, pmax = 1, deltaL = NULL, tauL = NULL, deltaR = NULL,
  tauR = NULL)
}
\arguments{
\item{transectDist}{The position at which to evaluate the cline equation.}

\item{decrease}{Is the cline decreasing in frequency? \code{TRUE} or
\code{FALSE}.}

\item{center}{The location of the cline center, in the same distance units as
\code{transectDist}. Numeric.}

\item{width}{The width of the cline, in the same distance units as
\code{transectDist}. Numeric, must be greater than 0.}

\item{pmin, pmax}{Optional. The minimum and maximum allele frequency values in
the tails of the cline. Default values are \code{0} and \code{1},
respectively. Must be between 0 and 1 (inclusive). Numeric.}

\item{deltaL, tauL}{Optional delta and tau parameters which describe the left
exponential introgression tail. Must supply both to generate a tail. Default is
\code{NULL} (no tails). Numeric. tauL must be between 0 and 1 (inclusive).}

\item{deltaR, tauR}{Optional delta and tau parameters which describe the right
exponential introgression tail. Must supply both to generate a tail. Default is
\code{NULL} (no tails). Numeric. tauR must be between 0 and 1 (inclusive).}
}
\value{
The result of evaluating a cline equation with the specified
parameters at the specified distance (that is, \eqn{p'} in the notation of the
equations above). A numeric vector of length 1.
}
\description{
A flexible function to evaluate hybrid zone genetic cline equations, with or
without introgression tails.
}
\details{
This function evaluates hybrid zone cline equations of the general form:

\deqn{p = (e^(4*((x-center))/(width))/((1+ e^(4*((x-center))/(wwidth)))}

Where \eqn{center} is the center of the cline, \eqn{width} is the width of
the cline, \eqn{x} is the distance along the transect, and \eqn{p} is the
expected frequency of the allele.

This function can also include introgression tails by specifying delta and
tau paramters. For the left introgression tail, when \eqn{x \le (center -
 \delta)}:

\deqn{p = 1/(1 + e^(4(\delta)/(w)))*\exp((4\tau(x-c+\delta)/width)/(1+e^(-4(\delta)/(w))))}

For the right introgression tail, when \eqn{x \ge (center + \delta)}:

\deqn{p = 1- 1/(1 + e^(4(\delta)/(w)))*\exp((-4\tau(x-c-\delta)/width)/(1+e^(-4(\delta)/(w))))}

The above equations all describe an increasing cline with minimum and
maximum values of 0 and 1. This equation can be rescaled for other
minima and maxima:

\deqn{p' = p_{min} + (p_{max} - p_{min})*p}

And for clines which decrease in frequency:

\deqn{p' = p_{min} + (p_{max} - p_{min})*(1-p)}

The function will automatically use the appropriate equation based on
arguments supplied by the user.
}
\examples{
\dontrun{
# Calculate the allele frequency at x = 100 for an increasing cline
# with a center at x = 125, a width of 40, and no tails.

general_cline_eqn(transectDist = 100, decrease = F,
                  center = 125, width = 40)

# Calculate the allele frequency at x = 84 for a decreasing cline
# with a center at 76, a width of 22, and a right tail
# with deltaR = 7.5, tauR = 0.45.

general_cline_eqn(transectDist = 84, decrease = TRUE,
                 center = 76, width = 22,
                 deltaR = 7.5, tauR = .45)
}

}
