% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_single_chain.R
\name{init_single_chain}
\alias{init_single_chain}
\title{Make the expression of init values for a single chain}
\usage{
init_single_chain(prior_file, tails = c("none", "left", "right",
  "mirror", "ind"))
}
\arguments{
\item{prior_file}{Path to the yaml file containing the prior specifications.}

\item{tails}{Which type of tails for the model: "none", "left", "right", "mirror", or "ind."}
}
\value{
An rlang expression of the initialization values for a single chain
}
\description{
This function it a bit complicated. It reads in the prior_file, and parses
both (1) the prior distribution for a parameter, and (2) the value(s) which
describe that distribution. It saves that info as an expression of the
appropriate R function to generate a single random value from that
distribution, for each parameter. Then, it assembles those parameters into a
list, only using the parameters needed for a certain model, specified by tails.
}
\details{
Currently supported distributions are: normal, uniform, exponential.

Used internally, in \link{make_init_list}.
}
\keyword{internal}
