% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_prior_file}
\alias{parse_prior_file}
\title{Parse the prior yaml file and check it}
\usage{
parse_prior_file(prior_file)
}
\arguments{
\item{prior_file}{filepath to the prior file}
}
\value{
a named list containing the specifed priors
}
\description{
DESCRIPTION TO BE ADDED
}
\details{
Used interanlly, in \link{create_cline_model} and \link{make_init_list}.
}
\keyword{internal}
