% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geno_cline.R
\name{plot_geno_cline}
\alias{plot_geno_cline}
\title{Plot thing}
\usage{
plot_geno_cline(stanfit, data, num.out = NULL, add.obs.freqs = F,
  point.col = "black", ...)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model results.}

\item{data}{The dataframe with your cline data.}

\item{num.out}{Optional, the number of x values (distances) at which to
evaluate the cline. By default, does twice the length of the cline, but you
can specify more or fewer. Too few may lead to a jagged-looking cline.}

\item{add.obs.freqs}{Should the observed allele frequencies at each site be
plotted? TRUE or FALSE, default is FALSE.}

\item{point.col}{The color to use for plotting the observed allele
frequencies. Default is black.}

\item{...}{Further graphical parameters to be passed to the base R plotting
functions, in particular graphical parameters (see \code{\link[graphics]{par}}).}
}
\value{
invisible(NULL)
}
\description{
Plot thing
}
