% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geno_cline.R
\name{plot_geno_cline}
\alias{plot_geno_cline}
\title{Plot a genetic cline}
\usage{
plot_geno_cline(stanfit, data, add.obs.freqs = F, point.col = "black",
  ...)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model
results.}

\item{data}{The dataframe with your cline data (ideally, the same data frame
that was used to generate the model fit).}

\item{add.obs.freqs}{Should the observed allele frequencies at each site be
plotted? TRUE or FALSE, default is FALSE.}

\item{point.col}{The color to use for plotting the observed allele
frequencies. Default is black.}

\item{...}{Further graphical parameters to be passed to the base R plotting
functions to customize the plot (see \code{\link[graphics]{par}}).}
}
\value{
invisible(NULL)
}
\description{
Makes a simple plot to visualize modeled genetic clines. The user supplies the stanfit
object containing the model fit and the dataframe with the original data, and
may also supply a number of optional arguments to customize the plot.
}
\details{
This plotting function is mostly a wrapper around
\code{\link{predict_cline}}. For greater customization of plots, users
are encouraged to use \code{\link{predict_cline}} to generate the x- and
y-coordinates for their fitted cline, and then graph those coordinates using
the plotting methods and packages of their choice (base plotting, lattice, or
ggplot2).
}
\examples{

\dontrun{

# Default plot with the cline only
plot_geno_cline(yourStanfit, data)

# Add points showing the empirical allele frequencies at
# each collecting site
plot_geno_cline(yourStanfit, data, add.obs.freq = T)

# Some plot customization
# Adding axis labels, titles, and changing the
# colors of the points and line.
plot_geno_cline(yourStanfit, data, add.obs.freq = T,
                main = "My cline",
                xlab = "distance",
                ylab = "allele frequency",
                point.col = "red",
                col = "blue")
}

}
