% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_geno_cline.R
\name{predict_geno_cline}
\alias{predict_geno_cline}
\title{Predict allele frequencies for a genetic cline}
\usage{
predict_geno_cline(stanfit, distance)
}
\arguments{
\item{stanfit}{A \code{\linkS4class{stanfit}} object holding your model results.}

\item{distance}{The x value(s) (distance along the transect) at which to
predict allele frequencies for the fitted cline. Must be a numeric vector.}
}
\value{
A data frame with two columns:
\itemize{
\item transectDist: the distances along the transect at which the cline
equation was evaluated. Will be the same values as were supplied in the
distance argument. Numeric.
\item p: the predict allele frequency at each distance, according to the
cline equation. Numeric.
}
}
\description{
Uses the cline parameters from your stanfit object (the mean of the posterior
distribtution of each parameter) to predict the allele frequency at the
distance(s) along the transect specified by the user.
}
\examples{
\dontrun{

# Predict at one distance
predict_geno_cline(yourStanfit, distance = 50)

# Predict at a range of distances
predict_geno_cline(yourStanfit, distance = 0:100)

}
}
