% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_prior_list.R
\name{prep_prior_list}
\alias{prep_prior_list}
\title{Prepare the list of prior values to be fed into Stan}
\usage{
prep_prior_list(prior_file)
}
\arguments{
\item{prior_file}{Path to the \code{.yaml} file containing the prior specifications.}
}
\value{
A list containing the values for the priors to be passed to Stan.
}
\description{
Parses the prior specifications provided in the \code{.yaml}
file, including checks that they are properly specified and
currently supported. Then, converts the prior specifications into
a named list to be passed to Stan.
}
\examples{
\dontrun{
prep_prior_list("path/to/priors.yaml")
}
}
\seealso{
\code{\link{fit_geno_cline}}
}
